Script behaviour and option parser.
Create the default behaviour for the script. Some adjustments may be
made but the basis should all be done.

Add an option parser to make parsing command line arguments easier. Work
on the parser is not complete but it is functional enough for the
example script.

Add an example script that only prints what the parser returns.