Translation foundations and no duplicate options
Add a function to translate text. Use this function on every sensible
output strings. For now, there is no detection of translation functions
and the translation only returns the same text.
Translation also replaces occurences of %(name)s tags with the
corresponding value in the optional second argument:

    _translate(
        "This %(thing)s is happening",
        array("thing"=>$something)
    )

Add functionality in optparse to refuse adding an option if it uses an
option text that was already used.