optparse: define callback signature
The callback signature of functions that get called upon processing of
an option was not well defined and only sufficient for the help option
callback.

Define the signature and pass all arguments to callback functions.
Change _optparse_display_help to match the new signature. Add
documentation for the signature so that users know how to define
callbacks.

The new signature has the following arguments:
    &$option : reference to the option being processed.
    $opt_text : the text that was used on the command line.
    $value : the value[1] that was given on the command line.
    &$parser : reference to the OptionParser object.

[1] : if multiple values were given to an option that required multiple
      values, this argument will be an array of all the values.

The -h option was broken by the last commit. It was requiring that a
value be given to the option. This is due to the fact that actions were
not implemented yet. Fix the option temporarily: add "nargs"=>0 to make
it require no value.

The help callback was displaying all options on the same line. Fix this
issue by appending "\n" at the end of each option display.