optparse: refactor names to comply to interface
Refactored the following names in order to make them identical to those
that correspond in Python's optparse:

* UnknownSettingsException => OptionError
* Option->nb_values => Option->nargs
* Option->help_text => Option->help
* OptionParser->_usage => OptionParser->usage
* OptionParser->_resolve_method => OptionParser->conflict_handler
* OptionParser->_usage => OptionParser->usage

Currently, set_usage handles replacement of %prog tag occurences so that
it is the program name. This is wrong in that the "usage" attribute
could be modified manually (by the user), and not replace %prog in this
context.

Fix this by replacing occurences of %prog only when printing usage.