optparse: implement OptionParser->print_help
Move code from _optparse_display_help to a new OptionParser method:
print_help. This makes it possible to display the help text even when
"add_help_option" is given a value of false.

Also, add translation for the "Option:" text and make it possible to
output the message to a different stream by passing the stream as an
argument.