optparse: support aggregates of short options
Currently, short options are taken as a whole. This means that it is not
possible to concatenate short options (say, "-vvvf bar" instead of "-v
-v -v -f bar").

Add support for aggregates by decomposing the option in a list of
characters. If an option takes at least one value, the rest of the
argument is taken as one of the needed value (so it is the last
aggregated option to be processed in this chunk). In clear, this means
that the following two use cases are similar: "-fbar" or "-f bar".