optparse: type checking + choices
Add the "type" argument to Option's constructor and enforce specified
type on the argument.

Add the "choices" argument to Option's constructor. Make "choice" the default
type if only the "choices" argument is given. Check that the given
values correspond to one of the choices.

also: change the default usage from "[arguments ...]" to "[options]".