optparse: comments and attibute visibility
A decision was made not to implement the OptionParser "option_list"
constructor attribute. Add a comment about that in the class comment.

Add a disambiguating comment about the default prog value.

Change visibility of Option attributes to protected so that subclasses
can play with them.
Method visibility was not yet evaluated and the are all still private.