optparse: epilog, metavar and standard_option_list
In OptionParser, add the "epilog" attribute so that a message can be
printed after the option list.

In Option, add the "metavar" attribute so that users can change what is
printed in the help message. Make it default to the uppercase value of
"dest".

Add a class attribute, in OptionParser, named standard_option_list that
defines the base list of options. This attribute can be used by
subclasses to feed a default list of options.